fix floats in the weightings.yaml causing loads of decimal places

+++ src/main/java/uk/ac/warwick/dcs/sherlock/engine/Registry.java
@@ -628,7 +628,7 @@ public class Registry implements IRegistry {
 					type.setWeighting(map.get(type.getIdentifier()).floatValue());
 				}
 				else {
-					map.put(type.getIdentifier(), (double) type.getWeighting());
+					map.put(type.getIdentifier(), type.getWeighting());
 				}
 			}
